Move IdentitySyncFlow into confidential-identities
